pyproject.toml
[tool.poetry]
name = "3d-holography"
version = "0.1.0"
description = ""
authors = ["shota_mizusaki <nrxg129@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.4"
opencv-python = "^4.9.0.80"
matplotlib = "^3.8.4"
torch = {version = "^2.3.0+cu121", source = "torch_cu121"}
torchvision = {version = "^0.18.0+cu121", source = "torch_cu121"}
torchaudio = {version = "^2.3.0+cu121", source = "torch_cu121"}
cupy-cuda12x = "^13.1.0"
h5py = "^3.11.0"
moviepy = "^1.0.3"
pandas = "^2.2.3"


[[tool.poetry.source]]
name = "torch_cu121"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
